ai_core/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.3.2
torch==2.5.1
requests==2.31.0
python-dotenv==1.0.0
grpcio==1.66.1
grpcio-tools==1.66.1
protobuf==5.27.3
redis==5.0.1
aiofiles==23.2.1
ibapi==9.81.1.post1
python-socketio==5.10.0
asyncio-mqtt==0.16.1